+++ src/types/calculated-model.tsx
@@ -1,28 +1,19 @@
-export interface IIntrinsicValue {
-  intrinsicValue: number;
-  acceptableBuyPrice: number;
-  differencePercentage: number;
-  belowIntrinsicValue: boolean;
-}
+import { IIntrinsicValue, IntrinsicValueDefault } from './intrinsic-value';
+import {
+  IPeterLynchValuationModel,
+  PeterLynchValuationModelDefault,
+} from './peter-lynch-valutation-model';
 
 export interface ICalculatedModel {
   grahamValutation: IIntrinsicValue;
   dcfValutation: IIntrinsicValue;
-  plValutation: number;
+  plValutation: IPeterLynchValuationModel;
+  belowIntrinsicValue: boolean;
 }
 
 export const defaultIntrinsicValue: ICalculatedModel = {
-  grahamValutation: {
-    intrinsicValue: 0,
-    acceptableBuyPrice: 0,
-    differencePercentage: 0,
-    belowIntrinsicValue: false,
-  },
-  dcfValutation: {
-    intrinsicValue: 0,
-    acceptableBuyPrice: 0,
-    differencePercentage: 0,
-    belowIntrinsicValue: false,
-  },
-  plValutation: 0,
+  grahamValutation: IntrinsicValueDefault,
+  dcfValutation: IntrinsicValueDefault,
+  plValutation: PeterLynchValuationModelDefault,
+  belowIntrinsicValue: false,
 };
