+++ src/types/intrinsic-value.tsx
@@ -1,0 +1,13 @@
+export interface IIntrinsicValue {
+  intrinsicValue: number;
+  acceptableBuyPrice: number;
+  differencePercentage: number;
+  belowIntrinsicValue: boolean;
+}
+
+export const IntrinsicValueDefault: IIntrinsicValue = {
+  intrinsicValue: 0,
+  acceptableBuyPrice: 0,
+  differencePercentage: 0,
+  belowIntrinsicValue: false,
+};
