+++ src/types/peter-lynch-valutation-model.tsx
@@ -1,0 +1,9 @@
+export interface IPeterLynchValuationModel {
+  plValutation: number;
+  belowIntrinsicValue: boolean;
+}
+
+export const PeterLynchValuationModelDefault: IPeterLynchValuationModel = {
+  plValutation: 0,
+  belowIntrinsicValue: false,
+};
